moved log in files
