+++ src/screens/index.ts
@@ -3,8 +3,8 @@ import InboxScreen from "./InboxScreenTab/InboxScreen"
 import ClientScreen from "./ClientTabScreen/ClientScreen"
 import ProtocolScreen from "./ProtocolTabScreen/ProtocolScreen"
 import CreateNewProtocol from "./ProtocolCreationScreens/CreateNewProtocol"
-import LoginScreen from "./LoginScreen"
-import SignUpScreen from "./SignUpScreen"
+import LoginScreen from "./LoginSignUpScreens/LoginScreen"
+import SignUpScreen from "./LoginSignUpScreens/SignUpScreen"
 
 export {
   DashboardScreen,
